1 file modified

+++ src/app/product-manager/product-manager.component.ts
@@ -11,7 +11,7 @@ import { ActivatedRoute, Router } from "@angular/router";
 export class ProductManagerComponent implements OnInit {
   selected: Product;
   products: Product[];
-  page =1 ;
+  page = 1;
   pageSize = 7;
   currentRate = 8;
 
@@ -19,7 +19,7 @@ export class ProductManagerComponent implements OnInit {
     private productService: ProductService,
     private route: ActivatedRoute,
     private router: Router
-  ) {}
+  ) { }
 
   ngOnInit(): void {
     this.getProducts();
@@ -34,10 +34,14 @@ export class ProductManagerComponent implements OnInit {
     });
   }
 
-  removeProduct(id){
-   this.productService.removeProduct(id).subscribe(response => {
-     this.products = this.products.filter(product => product.id != response.id);
-   })
+  removeProduct(id) {
+    let conf = confirm("bạn có chắc chắn muốn xóa không !");
+    if (conf) {
+      this.productService.removeProduct(id).subscribe(response => {
+        this.products = this.products.filter(product => product.id != response.id);
+      })
+    }
+
     // this.products = this.products.filter(product => product.id != id);
   }
 }
